Install DOOM via regular Emacs + manual commands
I was seeing issues related when adding config to "extraConfig" not being
picked up.

Overall, it was slightly harder to iterate my Emacs config. And each
small change required full re-installation of all packages, which was
slow.

I really want to use nix for everything but in this case it felt a bit
too painful; however, I am not super happy about the activation scripts
either. :)