Add (sly warning) to 'warning-suppress-types
Source: https://www.gnu.org/software/emacs/manual/html_node/elisp/Warning-Options.html

This list specifies which warning types should not be displayed immediately when they occur. Each element of the list should be a list of symbols. If an element of this list has the same elements as the first elements in a warning type, then the warning of that type will not be shown on display by popping the warnings buffer in some window (the warning will still be logged in the warnings buffer).